Fixes PATH to play nice with other buildpacks
This fixes the profile so it uses the PATH at time of load instead of
the PATH when the profile is compiled.

One of our customers (https://deadmanssnitch.com) ran into this where
our buildpack (https://github.com/deadmanssnitch/heroku-buildpack-dms)
wasn't available on the PATH because it was being overwritten.