Escape when writing to profile.d/freetds.sh
What?
=====

This updates the compile script to ensure environment variables exported
within profile.d/freetds.sh are escaped properly in the script rather
than evaluated when the script is run.

The previous behavior evaluated certain environment variables that were
intended to be prepended to, resulting in incorrect values in e.g.
`LIBRARY_PATH`.